refactor encoder and decoder into separate source files; improve func documentation
